+++ CHANGELOG.md
@@ -5,6 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.32.0] - 2025-05-02
+
+### Added
+
+- Compiler version is now shown during contract build
+- Added 'All Contracts' option to build wizard
+- Added function to build all tact contracts, required for rebuilding before tests
+
+### Changed
+
+- Made error of non-PascalCase contract names nicer
+
+### Fixed
+
+- `blueprint build --all` now exits with a non-zero exit code on failure
+
 ## [0.31.1] - 2025-04-24
 
 ### Fixed